pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kineo_tamarisk"
version = "0.1.0"
description = "Image-based salt cedar (tamarisk) identification."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Kineo Tamarisk Contributors" }
]
keywords = ["tamarisk", "salt-cedar", "classification", "pytorch"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "torch>=2.0",
  "torchvision>=0.15",
  "pillow>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
kineo-tamarisk-infer = "kineo_tamarisk.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["kineo_tamarisk*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
